services/file-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.4
